+++ src/backend/typeDefs/mutations/sitemap/index.js
@@ -1,7 +1,6 @@
 const { gql } = require('apollo-server-express')
 
 const updateSitemapTypeDefs = require('./update')
-const createSourceNodeTypeDefs = require('./createSourceNode')
 const updateTdgTimestampsTypeDefs = require('./updateTdgTimestamps')
 
 // need to create NodeInput of type Input, otherwise server fails:
@@ -26,6 +25,5 @@ const nodeInputTypedef = gql`
 module.exports = [
   nodeInputTypedef,
   updateSitemapTypeDefs,
-  createSourceNodeTypeDefs,
   updateTdgTimestampsTypeDefs,
 ]